js/events: hardcode path checking in cleanup task
Finding substring with currPath is faulty, as the
only time that currPath is updated is in the check
itself.

As such, a race condition might develop in which a
popup is reopened immediately after closing a popup.
In that case, currPath will not be '/events' but
the previous '/events/:link' and the new path will not
contain the substring, if the comparison was done with
currPath.

Signed-off-by: Syed Abdullah <contact@syedabdullah.net>